Move files to their final location (#707)
